agent/pyproject.toml
[project]
name = "mindrian-agent"
version = "0.1.0"
description = "Mindrian AI Agent powered by Agno"
requires-python = ">=3.11"
dependencies = [
    "agno[anthropic]>=2.3.14",
    "fastapi>=0.125.0",
    "httpx>=0.28.1",
    "psycopg[binary]>=3.3.2",
    "python-dotenv>=1.2.1",
    "ruff>=0.14.9",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.38.0",
]

[project.optional-dependencies]
dev = []

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
